fixed corrupted woff
