+++ src/app/globals.css
@@ -3,24 +3,13 @@
 @tailwind components;
 @tailwind utilities;
 
-/* Import Fonts */
-/* @import url("https://fonts.googleapis.com/css2?family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,100;1,300;1,400;1,500;1,700;1,900&family=RocknRoll+One&display=swap"); */
-
-/* @font-face {
-  font-family: "Blomberg";
-  src: url("/fonts/Blomberg-8MKKZ.woff2") format("woff2"),
-    url("/fonts/Blomberg-8MKKZ.woff") format("woff");
-  font-weight: normal;
-  font-style: normal;
-} */
-
 /* Global Styles */
 html {
   scroll-behavior: smooth;
 }
 
 body {
-  /* font-family: "RocknRoll One", sans-serif; */
+  font-family: var(--font-rocknRoll-one), sans-serif;
   font-weight: 100;
   font-style: normal;
   color: #213445;
@@ -31,7 +20,7 @@ body {
 /* Typography */
 h1,
 h3 {
-  /* font-family: "Blomberg", system-ui; */
+  font-family: "Blomberg", sans-serif;
   font-weight: normal;
   font-style: normal;
 }
