+++ src/app/layout.tsx
@@ -1,4 +1,27 @@
 import type { Metadata } from "next";
+import { RocknRoll_One, Roboto } from "next/font/google";
+import localFont from "next/font/local";
+
+const blombergFont = localFont({
+  src: "./fonts/BLOMBERG2.woff2",
+  weight: "normal",
+  style: "normal",
+  variable: "--font-blomberg",
+});
+
+const rocknRollOne = RocknRoll_One({
+  weight: "400",
+  subsets: ["latin"],
+  variable: "--font-rocknRoll-one",
+});
+
+const roboto = Roboto({
+  weight: ["300", "400", "700"],
+  subsets: ["latin"],
+  variable: "--font-roboto",
+});
+
+import "./globals.css";
 
 export const metadata: Metadata = {
   title: "Veronica Coulombe - Front-End Web Developer",
@@ -15,7 +38,10 @@ export default function RootLayout({
   children: React.ReactNode;
 }>) {
   return (
-    <html lang="en">
+    <html
+      lang="en"
+      className={`${rocknRollOne.variable} ${roboto.variable} ${blombergFont.variable}`}
+    >
       <body className="base-100">{children}</body>
     </html>
   );
